+++ src/helpers/extract-fiffteen-words.ts
